Refactoring. Move IPrinter.
