+++ Task_02/Printers/Abstract/IPrinter.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Task_02.Statistics
+namespace Task_02.Printers.Abstract
 {
     internal interface IPrinter
     {
