+++ Task_02/Printers/Abstract/NumberOfMatches/NumberOfMatchesPrinter.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using Task_02.Statistics;
 
 namespace Task_02.Printers.Abstract.NumberOfMatches
 {
