+++ Task_02/Statistics/AbstractStaticticProvider.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Task_02.Printers.Abstract;
 
 namespace Task_02.Statistics
 {
