+++ Task_02/Statistics/HeroUsagesStatisticWinStreakProvider.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using Task_02.Printers;
+using Task_02.Printers.Abstract;
 
 namespace Task_02.Statistics
 {
